WIP driver plugin implementation
