+++ boilingcore/boilingcore.go
@@ -14,7 +14,6 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/vattle/sqlboiler/bdb"
-	"github.com/vattle/sqlboiler/bdb/drivers"
 	"github.com/vattle/sqlboiler/queries"
 	"github.com/vattle/sqlboiler/strmangle"
 )
@@ -256,44 +255,6 @@ func getBasePath(baseDirConfig string) (string, error) {
 	}
 
 	return os.Getwd()
-}
-
-// initDriver attempts to set the state Interface based off the passed in
-// driver flag value. If an invalid flag string is provided an error is returned.
-func (s *State) initDriver(driverName string) error {
-	// Create a driver based off driver flag
-	switch driverName {
-	case "postgres":
-		s.Driver = drivers.NewPostgresDriver(
-			s.Config.Postgres.User,
-			s.Config.Postgres.Pass,
-			s.Config.Postgres.DBName,
-			s.Config.Postgres.Host,
-			s.Config.Postgres.Port,
-			s.Config.Postgres.SSLMode,
-		)
-	case "mysql":
-		s.Driver = drivers.NewMySQLDriver(
-			s.Config.MySQL.User,
-			s.Config.MySQL.Pass,
-			s.Config.MySQL.DBName,
-			s.Config.MySQL.Host,
-			s.Config.MySQL.Port,
-			s.Config.MySQL.SSLMode,
-		)
-	case "mock":
-		s.Driver = &drivers.MockDriver{}
-	}
-
-	if s.Driver == nil {
-		return errors.New("An invalid driver name was provided")
-	}
-
-	s.Dialect.LQ = s.Driver.LeftQuote()
-	s.Dialect.RQ = s.Driver.RightQuote()
-	s.Dialect.IndexPlaceholders = s.Driver.IndexPlaceholders()
-
-	return nil
 }
 
 // initTables retrieves all "public" schema table names from the database.
